+++ src/Paket/AssemblyInfo.fs
@@ -5,10 +5,10 @@ open System.Reflection
 [<assembly: AssemblyProductAttribute("Paket")>]
 [<assembly: AssemblyCompanyAttribute("Paket team")>]
 [<assembly: AssemblyDescriptionAttribute("A package dependency manager for .NET with support for NuGet packages and GitHub repositories.")>]
-[<assembly: AssemblyVersionAttribute("2.3.6")>]
-[<assembly: AssemblyFileVersionAttribute("2.3.6")>]
-[<assembly: AssemblyInformationalVersionAttribute("2.3.6")>]
+[<assembly: AssemblyVersionAttribute("2.5.2")>]
+[<assembly: AssemblyFileVersionAttribute("2.5.2")>]
+[<assembly: AssemblyInformationalVersionAttribute("2.5.2")>]
 do ()
 
 module internal AssemblyVersionInformation =
-    let [<Literal>] Version = "2.3.6"
+    let [<Literal>] Version = "2.5.2"