+++ tests/Paket.Tests/DependencySetSpecs.fs
@@ -18,14 +18,11 @@ let ``should optimize 2 restriction set with only exactly``() =
          PackageName("P2"), VersionRequirement.AllReleases, [FrameworkRestriction.Exactly (DotNetFramework(FrameworkVersion.V3_5))]]
 
     let expected =
-        [PackageName("P1"), VersionRequirement.AllReleases, 
-            [FrameworkRestriction.Exactly (DotNetFramework(FrameworkVersion.V3_5))
-             FrameworkRestriction.Exactly (DotNetFramework(FrameworkVersion.V4))
-             FrameworkRestriction.AtLeast (DotNetFramework(FrameworkVersion.V4_5))]
-         PackageName("P2"), VersionRequirement.AllReleases, [FrameworkRestriction.Between (DotNetFramework(FrameworkVersion.V3_5),DotNetFramework(FrameworkVersion.V4))]]
+        [PackageName("P1"), VersionRequirement.AllReleases, [FrameworkRestriction.AtLeast (DotNetFramework(FrameworkVersion.V3_5))]
+         PackageName("P2"), VersionRequirement.AllReleases, [FrameworkRestriction.Exactly (DotNetFramework(FrameworkVersion.V3_5))]]
 
     original
-    |> optimizeRestrictions
+    |> optimizeDependencies
     |> shouldEqual expected
 
 [<Test>]
@@ -37,14 +34,11 @@ let ``should optimize 2 restriction set with only exactly and client framework``
          PackageName("P2"), VersionRequirement.AllReleases, [FrameworkRestriction.Exactly (DotNetFramework(FrameworkVersion.V3_5))]]
 
     let expected =
-        [PackageName("P1"), VersionRequirement.AllReleases, 
-            [FrameworkRestriction.Exactly (DotNetFramework(FrameworkVersion.V3_5))
-             FrameworkRestriction.Exactly (DotNetFramework(FrameworkVersion.V4_Client))
-             FrameworkRestriction.AtLeast (DotNetFramework(FrameworkVersion.V4_5))]
-         PackageName("P2"), VersionRequirement.AllReleases, [FrameworkRestriction.Between (DotNetFramework(FrameworkVersion.V3_5),DotNetFramework(FrameworkVersion.V4_Client))]]
+        [PackageName("P1"), VersionRequirement.AllReleases, [FrameworkRestriction.AtLeast (DotNetFramework(FrameworkVersion.V3_5))]
+         PackageName("P2"), VersionRequirement.AllReleases, [FrameworkRestriction.Exactly (DotNetFramework(FrameworkVersion.V3_5))]]
 
     original
-    |> optimizeRestrictions
+    |> optimizeDependencies
     |> shouldEqual expected
 
 [<Test>]
@@ -56,14 +50,11 @@ let ``should optimize 2 restriction sets with between``() =
          PackageName("P2"), VersionRequirement.AllReleases, [FrameworkRestriction.Between (DotNetFramework(FrameworkVersion.V3_5),DotNetFramework(FrameworkVersion.V4_Client))]]
 
     let expected =
-        [PackageName("P1"), VersionRequirement.AllReleases, 
-            [FrameworkRestriction.Exactly (DotNetFramework(FrameworkVersion.V3_5))
-             FrameworkRestriction.Exactly (DotNetFramework(FrameworkVersion.V4))
-             FrameworkRestriction.AtLeast (DotNetFramework(FrameworkVersion.V4_5))]
+        [PackageName("P1"), VersionRequirement.AllReleases, [FrameworkRestriction.AtLeast (DotNetFramework(FrameworkVersion.V3_5))]
          PackageName("P2"), VersionRequirement.AllReleases, [FrameworkRestriction.Between (DotNetFramework(FrameworkVersion.V3_5),DotNetFramework(FrameworkVersion.V4_Client))]]
 
     original
-    |> optimizeRestrictions
+    |> optimizeDependencies
     |> shouldEqual expected
 
 [<Test>]
@@ -132,4 +123,25 @@ let ``filtered with AtLeast restriction should filter non-matching``() =
 
     original
     |> DependencySetFilter.filterByRestrictions [FrameworkRestriction.AtLeast (DotNetFramework(FrameworkVersion.V4))]
+    |> shouldEqual expected
+
+[<Test>]
+let ``should optimize ZendeskApi_v2 ``() = 
+    let original =
+        [PackageName("Newtonsoft.Json"), VersionRequirement.AllReleases, [FrameworkRestriction.Exactly (DotNetFramework(FrameworkVersion.V3_5))]
+         PackageName("Newtonsoft.Json"), VersionRequirement.AllReleases, [FrameworkRestriction.Exactly (DotNetFramework(FrameworkVersion.V4))]
+         PackageName("AsyncCTP"), VersionRequirement.AllReleases, [FrameworkRestriction.Exactly (DotNetFramework(FrameworkVersion.V4))]
+         PackageName("Newtonsoft.Json"), VersionRequirement.AllReleases, [FrameworkRestriction.Exactly (DotNetFramework(FrameworkVersion.V4_5))]
+         PackageName("Newtonsoft.Json"), VersionRequirement.AllReleases, [FrameworkRestriction.Portable "portable-net45+sl40+wp71+win80"]
+         PackageName("Microsoft.Bcl.Async"), VersionRequirement.AllReleases, [FrameworkRestriction.Portable "portable-net45+sl40+wp71+win80"]]
+
+    let expected =
+        [PackageName("Newtonsoft.Json"), VersionRequirement.AllReleases, 
+            [FrameworkRestriction.Portable "portable-net45+sl40+wp71+win80"
+             FrameworkRestriction.AtLeast (DotNetFramework(FrameworkVersion.V3_5))]
+         PackageName("AsyncCTP"), VersionRequirement.AllReleases, [FrameworkRestriction.Exactly (DotNetFramework(FrameworkVersion.V4))]
+         PackageName("Microsoft.Bcl.Async"), VersionRequirement.AllReleases, [FrameworkRestriction.Portable "portable-net45+sl40+wp71+win80"]]
+
+    original
+    |> optimizeDependencies
     |> shouldEqual expected